modules/connection-grpc/requirements.txt
grpcio
grpcio-tools
protobuf
psycopg2-binary
SQLAlchemy
geoalchemy2
Shapely==1.7.1
flask_sqlalchemy